fix(in_port, network id): Update the in_port and neutron network id
Updates in Guided solution (page 2):

Sample output in section 2:

In the ovs-ofctl output example, change the in_port from 6 to 8 to match the id used in the ensuing in-depth open flow analysis.

Description in Section 5:

In the description of the ovn-sbctl output, update the neutron network id to match the one used throughout the document.